--- a/bulk_query_gen/influxdb/influx_iot_common.go
+++ b/bulk_query_gen/influxdb/influx_iot_common.go
@@ -124,3 +124,24 @@ func (d *InfluxIot) LightLevelEightHours(qi bulkQuerygen.Query) {
 	q := qi.(*bulkQuerygen.HTTPQuery)
 	d.getHttpQuery(humanLabel, "n/a", query, q)
 }
+
+func (d *InfluxIot) IotSortedPivot(qi bulkQuerygen.Query) {
+	var query string
+	if d.language == InfluxQL {
+		query = fmt.Sprintf(`SELECT * WHERE time > '%s' AND time < '%s'`,
+			d.AllInterval.StartString(),
+			d.AllInterval.EndString())
+	} else {
+		query = fmt.Sprintf(`from(bucket: "%s" `+
+			`|> range(start: %s, stop: %s) `+
+			`|> filter(fn: (r) => r._measurememt == "air_quality_room") `+
+			`|> pivot(rowKey:["_time"], columnKey:["co2_level"])`, // todo do i need value key?
+			d.DatabaseName,
+			d.AllInterval.StartString(),
+			d.AllInterval.EndString())
+	}
+
+	humanLabel := fmt.Sprintf(`InfluxDB (%s) Sorted Pivot`, d.language)
+	q := qi.(*bulkQuerygen.HTTPQuery)
+	d.getHttpQuery(humanLabel, d.AllInterval.StartString(), query, q)
+}