--- a/bulk_query_gen/influxdb/influx_iot_common.go
+++ b/bulk_query_gen/influxdb/influx_iot_common.go
@@ -132,10 +132,10 @@ func (d *InfluxIot) IotSortedPivot(qi bulkQuerygen.Query) {
 			d.AllInterval.StartString(),
 			d.AllInterval.EndString())
 	} else {
-		query = fmt.Sprintf(`from(bucket: "%s" `+
+		query = fmt.Sprintf(`from(bucket: "%s") `+
 			`|> range(start: %s, stop: %s) `+
 			`|> filter(fn: (r) => r._measurememt == "air_quality_room") `+
-			`|> pivot(rowKey:["_time"], columnKey:["co2_level"])`, // todo do i need value key?
+			`|> pivot(rowKey:["_time"], columnKey:["_field"], valueKey:["_value"])`,
 			d.DatabaseName,
 			d.AllInterval.StartString(),
 			d.AllInterval.EndString())