--- a/bulk_query_gen/influxdb/influx_iot_sorted_pivot.go
+++ b/bulk_query_gen/influxdb/influx_iot_sorted_pivot.go
@@ -9,24 +9,27 @@ import (
 // on Flux pivot function
 type InfluxIotSortedPivot struct {
 	InfluxIot
+	interval time.Duration
 }
 
 func NewInfluxQLIotSortedPivot(dbConfig bulkQuerygen.DatabaseConfig, queriesFullRange bulkQuerygen.TimeInterval, queryInterval time.Duration, scaleVar int) bulkQuerygen.QueryGenerator {
 	underlying := NewInfluxIotCommon(InfluxQL, dbConfig, queriesFullRange, queryInterval, scaleVar).(*InfluxIot)
 	return &InfluxIotSortedPivot{
 		InfluxIot: *underlying,
+		interval: queryInterval,
 	}
 }
 
 func NewFluxIotSortedPivot(dbConfig bulkQuerygen.DatabaseConfig, queriesFullRange bulkQuerygen.TimeInterval, queryInterval time.Duration, scaleVar int) bulkQuerygen.QueryGenerator {
 	underlying := NewInfluxIotCommon(Flux, dbConfig, queriesFullRange, queryInterval, scaleVar).(*InfluxIot)
-	return &InfluxIotAggregateKeep{
+	return &InfluxIotSortedPivot{
 		InfluxIot: *underlying,
+		interval: queryInterval,
 	}
 }
 
 func (d *InfluxIotSortedPivot) Dispatch(i int) bulkQuerygen.Query {
 	q := bulkQuerygen.NewHTTPQuery()
-	d.IotSortedPivot(q)
+	d.IotSortedPivot(q, d.interval)
 	return q
 }