Perf test for https://github.com/influxdata/influxdb/issues/22016.

The performance data I gathered:
||   max   |   mean  |
|----------------|:-------:|:-------:|
| InfluxQL (1.8) | 4.29ms | 0.26ms |
| Flux (2.x) | 19.39ms | 14.03ms |

This data is not completely fair, since the `pivot()` function does not exist in InfluxQL as stated [here](https://docs.influxdata.com/influxdb/v1.8/flux/flux-vs-influxql/#pivot), but since InfluxQL will pivot automatically, it should still be a valid test.