@DStrand1 2021-08-11
Addressed requested changes, and updated the benchmarking results table in the opening comment.