@DStrand1 2021-08-11
Not really sure why these types of errors are even running at all in the query running but, fixed them and updated the table once more.